pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pspm"
version = "0.1.0"
description = "Reproducible PM/PS metrics with diffusion maps for speech separation research."
authors = [{name = "YOUR NAME", email = "YOUR.EMAIL@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["speech", "audio", "diffusion maps", "perceptual metrics", "PM", "PS"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "librosa>=0.10",
  "pyloudnorm>=0.1.1",
  "torch>=2.1",
  "transformers>=4.40",
  "tqdm>=4.66",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "black>=24.3",
  "ruff>=0.4.0",
  "mypy>=1.10",
  "pytest>=8.0",
  "pre-commit>=3.7"
]

[project.scripts]
pspm-run = "pspm.cli:main"
